chore(zset): move skiplist impl to another file
